requirements.txt
fastapi==0.109.0
uvicorn
python-multipart
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
accelerate==0.26.1
optimum==1.16.1
insanely-fast-whisper==0.0.13
pydantic-settings==2.2.1
boto3
python-dotenv
sqlalchemy==2.0.29
asyncpg==0.29.0